Re-organise Ethereum related modules in ethereum mod
