+++ src/ethereum.rs
@@ -1,0 +1,6 @@
+pub mod dai;
+mod geth;
+mod wallet;
+
+pub use geth::Client;
+pub use wallet::Wallet;
