+++ src/lib.rs
@@ -16,10 +16,8 @@
 
 pub mod bitcoin;
 pub mod config;
-pub mod dai;
-pub mod ethereum_wallet;
+pub mod ethereum;
 pub mod float_maths;
-pub mod geth;
 pub mod history;
 pub mod jsonrpc;
 pub mod maker;
