+++ src/main.rs
@@ -11,7 +11,8 @@ use futures_timer::Delay;
 use nectar::{
     bitcoin, config,
     config::{settings, Settings},
-    dai, ethereum_wallet,
+    ethereum,
+    ethereum::dai,
     maker::{Free, PublishOrders, TakeRequestDecision},
     mid_market_rate::get_btc_dai_mid_market_rate,
     network::{self, Nectar, Orderbook, Taker},
@@ -27,7 +28,7 @@ const ENSURED_CONSUME_ZERO_BUFFER: usize = 0;
 
 async fn init_maker(
     bitcoin_wallet: Arc<bitcoin::Wallet>,
-    ethereum_wallet: Arc<ethereum_wallet::Wallet>,
+    ethereum_wallet: Arc<ethereum::Wallet>,
     maker_settings: settings::Maker,
 ) -> Maker {
     let initial_btc_balance = bitcoin_wallet.balance().await;
@@ -111,7 +112,7 @@ fn init_bitcoin_balance_updates(
 
 fn init_dai_balance_updates(
     update_interval: Duration,
-    wallet: Arc<ethereum_wallet::Wallet>,
+    wallet: Arc<ethereum::Wallet>,
 ) -> (
     impl Future<Output = comit::Never> + Send,
     Receiver<anyhow::Result<dai::Amount>>,
@@ -137,7 +138,7 @@ fn init_dai_balance_updates(
 async fn execute_swap(
     db: Arc<Database>,
     bitcoin_wallet: Arc<bitcoin::Wallet>,
-    ethereum_wallet: Arc<ethereum_wallet::Wallet>,
+    ethereum_wallet: Arc<ethereum::Wallet>,
     bitcoin_connector: Arc<comit::btsieve::bitcoin::BitcoindConnector>,
     ethereum_connector: Arc<comit::btsieve::ethereum::Web3Connector>,
     swap_execution_finished_sender: Sender<FinishedSwap>,
@@ -213,7 +214,7 @@ fn handle_network_event(
     swarm: &mut libp2p::Swarm<Nectar>,
     db: Arc<Database>,
     bitcoin_wallet: Arc<bitcoin::Wallet>,
-    ethereum_wallet: Arc<ethereum_wallet::Wallet>,
+    ethereum_wallet: Arc<ethereum::Wallet>,
     bitcoin_connector: Arc<comit::btsieve::bitcoin::BitcoindConnector>,
     ethereum_connector: Arc<comit::btsieve::ethereum::Web3Connector>,
     sender: Sender<FinishedSwap>,
@@ -390,7 +391,7 @@ async fn main() {
     .expect("can initialise bitcoin wallet");
     let bitcoin_wallet = Arc::new(bitcoin_wallet);
     let ethereum_wallet =
-        ethereum_wallet::Wallet::new(seed, settings.ethereum.node_url, dai_contract_addr)
+        ethereum::Wallet::new(seed, settings.ethereum.node_url, dai_contract_addr)
             .expect("can initialise ethereum wallet");
     let ethereum_wallet = Arc::new(ethereum_wallet);
     let maker = init_maker(bitcoin_wallet, ethereum_wallet, settings.maker).await;
@@ -473,7 +474,7 @@ async fn main() {
 fn respawn_swaps(
     db: Arc<Database>,
     bitcoin_wallet: Arc<bitcoin::Wallet>,
-    ethereum_wallet: Arc<ethereum_wallet::Wallet>,
+    ethereum_wallet: Arc<ethereum::Wallet>,
     bitcoin_connector: Arc<comit::btsieve::bitcoin::BitcoindConnector>,
     ethereum_connector: Arc<comit::btsieve::ethereum::Web3Connector>,
     swap_execution_finished_sender: Sender<FinishedSwap>,
